Fix warnings from XS.xs when compiled with -Wall
There was also a real bug in the in_pad() sub. It was checking if the pad name
was SVf_OK in a way that would always think it was okay because it was missing
some parens around a negation operator.

It's not clear to me what type of code would trigger this check, however, so I
don't know how to test this. None of the existing tests were exercising this
condition either, as removing it entirely does not change whether tests pass
or fail.